pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["Cácio Lucas <cclucas060901@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "django-wkhtmltopdf"
dj-database-url = "^1.0.0"
python-decouple = "^3.6"
psycopg2 = "^2.9.3"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"
django-extensions = "^3.2.1"
django-filter = "^22.1"
django-storages = "^1.13.1"
boto3 = "^1.26.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
